requirements.txt
flask==3.0.2
openai==1.8.2
pandas==2.2.1
fpdf2==2.7.8
requests==2.31.0
python-dotenv==1.0.1
autogen==0.9.1.post0
SQLAlchemy==2.0.28
python-dateutil==2.8.2 